set position of emojiPicker
https://youtrack.weseek.co.jp/issue/GW-7750
- Clear emojiSearchText on modal close
- Move getEmojiTranslation to emojiPickerHelper
- Disable fade effect of Modal
- Remove focus on editor after modal opened
- Change emojiPicker position to fixed
- Add space after emoji-colons
