+++ packages/app/src/components/PageEditor/EmojiPicker.tsx
@@ -1,12 +1,11 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 
 import { Picker } from 'emoji-mart';
-import i18n from 'i18next';
 import { Modal } from 'reactstrap';
 
 import { isDarkMode } from '~/client/util/color-scheme';
 
-import EmojiPickerHelper from './EmojiPickerHelper';
+import EmojiPickerHelper, { getEmojiTranslation } from './EmojiPickerHelper';
 
 type Props = {
   onClose: () => void,
@@ -22,7 +21,6 @@ const EmojiPicker: FC<Props> = (props: Props) => {
     onClose, emojiSearchText, emojiPickerHelper, isOpen,
   } = props;
 
-  const [style, setStyle] = useState({});
   // Set search emoji input and trigger search
   const searchEmoji = () => {
     if (emojiSearchText !== null) {
@@ -32,8 +30,6 @@ const EmojiPicker: FC<Props> = (props: Props) => {
       const event = new Event('input', { bubbles: true });
       input.dispatchEvent(event);
       input.focus();
-      const emojiPickerHeight = window.document.querySelector('[id^="emoji-mart-search"]')?.clientHeight;
-      setStyle(emojiPickerHelper.getCursorCoords(emojiPickerHeight));
     }
   };
 
@@ -48,60 +44,21 @@ const EmojiPicker: FC<Props> = (props: Props) => {
   };
 
 
-  const getEmojiTranslation = () => {
-
-    const categories = {};
-    [
-      'search',
-      'recent',
-      'smileys',
-      'people',
-      'nature',
-      'foods',
-      'activity',
-      'places',
-      'objects',
-      'symbols',
-      'flags',
-      'custom',
-    ].forEach((category) => {
-      categories[category] = i18n.t(`emoji.categories.${category}`);
-    });
-
-    const skintones = {};
-    (Array.from(Array(6).keys())).forEach((tone) => {
-      skintones[tone + 1] = i18n.t(`emoji.skintones.${tone + 1}`);
-    });
-
-    const translation = {
-      search: i18n.t('emoji.search'),
-      clear: i18n.t('emoji.clear'),
-      notfound: i18n.t('emoji.notfound'),
-      skintext: i18n.t('emoji.skintext'),
-      categories,
-      categorieslabel: i18n.t('emoji.categorieslabel'),
-      skintones,
-      title: i18n.t('emoji.title'),
-    };
-
-    return translation;
-  };
-
   const translation = getEmojiTranslation();
   const theme = isDarkMode() ? 'dark' : 'light';
 
-  return Object.keys(style).length !== 0 ? (
-    <Modal isOpen={isOpen} toggle={onClose} onOpened={searchEmoji}>
+  return (
+    <Modal isOpen={isOpen} toggle={onClose} onOpened={searchEmoji} fade={false}>
       <Picker
         onSelect={selectEmoji}
         i18n={translation}
         title={translation.title}
         emojiTooltip
-        style={{ position: 'absolute' }}
+        style={emojiPickerHelper.setStyle()}
         theme={theme}
       />
     </Modal>
-  ) : <></>;
+  );
 };
 
 export default EmojiPicker;
